% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_quantities_RIE_separate.R
\name{df_quantities_RIE_separate}
\alias{df_quantities_RIE_separate}
\title{Calculates or estimates RIE for raw and worked ivory seizures}
\usage{
df_quantities_RIE_separate(year.from = 1900, year.to = 2100,
  statusMin = 3, reg.model = "wt est.Rdata")
}
\arguments{
\item{year.from}{Starting year}

\item{year.to}{Ending year}

\item{statusMin}{Minimum status of seizure record to include in summary (usually 3)}

\item{reg.model}{.Rdata file that contains parameters for estimating weights from pieces}
}
\value{
A dataframe which lists for each seizure their:
  id, country of discovery, year & ivory quantities (weight and pieces)
  including RIE separately for raw and worked. \cr Note that seizures which are
  all non-ivory or are ivory seizuress with no information about quantities are excluded.
}
\description{
Produces a list of all seizures and their RIE
for raw and worked ivory. For seizures where only the number of pieces
is provided the weight is estimated.
}
\details{
This function interacts with the ETIS database, to extract seizure information.
It also requires the wt est.Rdata file that was created by the last COP analysis to estimate the
weight of seizures where only the number of pieces if provided. Note that since previous ETIS analyses,
the model estimates weights from log(pieces + 1).

The main use of this function is to contribute to the calculation
of two new database tables for law enforcement ratio and trade flow accounting.

This function is also used in the ETIS database to produce various summary tables
}
\examples{
year.from <- 1900
year.to <- 2100
statusMin <- 3

df.1 <- df_quantities_RIE_separate(year.from = year.from, year.to = year.to,
statusMin = statusMin, reg.model = 'wt est.Rdata')
}
\seealso{
The dataframe is used in \code{\link{mult_ctries}}, and
\code{\link{inout_tables_LE_TF_prep}}. These are then processed using
\code{\link{inout_tables_LE_TF_final}} to give the final set of outputs.
}
