% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_tables_final.R
\name{inout_tables_final}
\alias{inout_tables_final}
\title{Adjusts seizures and weights out given multiple countries of origin.}
\usage{
inout_tables_final(year.from = 1900, year.to = 2100, statusMin = 3,
  ctry_dest_included, size.min = 0, which.out = 1:3, mult.data,
  inout.data)
}
\arguments{
\item{year.from}{Starting year}

\item{year.to}{Ending year}

\item{statusMin}{Minimum status of seizure record to include in summary (usually 3)}

\item{ctry_dest_included}{true if ctry of dest included in computing szs out
TRUE for trade flows and FALSE for LE ratio accounting. If this is different to
the options for df.inout then the function aborts}

\item{size.min}{This is by default set to zero but is useful when wanting list of seizures
used in cluster analysis when a size cutoff is used}

\item{which.out}{which outputs to include. Denotes which of the outputs to be
produced. If it includes 2 then year.from = 1900, year.to = 2100 and size.min = 0.}

\item{mult}{List created using mult_ctries() with countries of
discovery, transit and destination for seizures with multiple countries of origin}

\item{df.inout}{List created using inout_tables_LE_TF_prep with  list
of seizures and weights in and out uncorrected for multiple countries of origin}
}
\value{
This writes files to three different places. Each gives
seizures and weights in and out for each country and year specified:
\enumerate{
\item csv file: NAME.csv
\item database table: szwt_NAME - requires size.min = 0, year.from = 1900, year.to = 2100
\item Excel workbook: NAME.xlsx with separate sheets for:
    sz in, sz out, wt in, wt out, sz flow, wt flow
   }
The file names depend on the accounting method that is required and the years included.
So file name might be LE_1900_2100.csv or TF_1900_2100.csv.
The start of the file name relates to:
\tabular{rl}{
LE \tab for LE ratio accounting - the country of destination excluded \cr
TF \tab if for trade flow accounting - the country of destination included \cr
}
The years are set as this large range when writing to the database#
}
\description{
Computes corrected number of seizures and weights in and out after adjusting for
multiple countries of origin. Outputs can be saved as a csv file, or into the
database or as an Excel Workbook.
}
\details{
The output from this function is a list of seizures and weights in and out.
The user can choose which outputs are required.

This function interacts with the ETIS database and the list of relevant seizures
stored in df.1 from df_quantities_RIE_separate.fn.

The main use of this function is to contribute to the calculation
of two new database tables for law enforcement ratio and trade flow accounting.

This function makes an adjustment for seizure 109462. For this seizure there is
a country of transit listed that falls after the country of discovery. For that country,
the seizure needs to be included in trade flow but not LE ratio accounting.
}
\examples{
year.from <- 1900
year.to <- 2100
statusMin <- 3
which.out <- 1:3

df.1 <- df_quantities_RIE_separate(year.from = year.from, year.to = year.to,
statusMin = statusMin, reg.model = 'wt est.Rdata')

mult.dat <- mult_ctries(year.from = 1900, year.to = 2100, statusMin = 3, df.1 = df.1)

ctry_dest_included <- FALSE

df.inout <- inout_tables_prep(year.from = year.from,year.to = year.to,
statusMin = statusMin,ctry_dest_included = ctry_dest_included, df.1 = df.1)

inout_tables_final(year.from = year.from, year.to = year.to,
statusMin = statusMin, ctry_dest_included = ctry_dest_included,
size.min = size.min, which.out = which.out, mult.data = mult.dat,
inout.data = df.inout)
}
\seealso{
\code{\link{df_quantities_RIE_separate}},
\code{\link{mult_ctries}},
\code{\link{inout_tables_prep}}
}
