% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout_tables_prep.R
\name{inout_tables_prep}
\alias{inout_tables_prep}
\title{Prepares data for  szs and wgts in and out}
\usage{
inout_tables_prep(year.from = 1900, year.to = 2100, statusMin = 3,
  ctry_dest_included, df.1 = df.1)
}
\arguments{
\item{year.from}{Starting year}

\item{year.to}{Ending year}

\item{statusMin}{Minimum status of seizure record to include in summary (usually 3)}

\item{ctry_dest_included}{true if ctry of dest included in computing szs out
TRUE for trade flows and FALSE for LE ratio accounting}

\item{df.1}{Dataframe with list of seizure records to be considered.
Created using df_quantities_RIE_separate.fn}
}
\value{
a list with two items
df_inout A dataframe listing seizures and weights in and out for each country in each year
ctry_dest_included Flag denoting whether or not country of destination included
}
\description{
The output from this function is a list of seizures and weights in and out
that has not been adjusted for multiple countries of origin
This function interacts with the ETIS database and the list of relevant seizures
stored in df.1 from df_quantities_RIE_separate.fn
The main use of this function is to contribute to the calculation
of two new database tables for law enforcement ratio and trade flow accounting.
}
\details{
This program also deals with seizure 109462.
One of the countries of transit that is listed occurs after the country of discovery.
For that country, the seizure needs to be included in trade flow accounting and excluded from LE ratio accounting
}
\examples{
year.from <- 1900
year.to <- 2100
statusMin <- 3

df.1 <- df_quantities_RIE_separate(year.from = year.from, year.to = year.to,
statusMin = statusMin, reg.model = 'wt est.Rdata')

mult.dat <- mult_ctries(year.from = 1900, year.to = 2100, statusMin = 3, df.1 = df.1)

ctry_dest_included <- FALSE

df.inout <- inout_tables_prep(year.from = year.from, year.to = year.to,
statusMin = statusMin, ctry_dest_included = ctry_dest_included,
df.1 = df.1)
}
\seealso{
\code{\link{df_quantities_RIE_separate}},
\code{\link{mult_ctries}},
\code{\link{inout_tables_final}}
}
