% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_ctries.R
\name{mult_ctries}
\alias{mult_ctries}
\title{Identifies seizures with multiple countries of origin}
\usage{
mult_ctries(year.from = 1900, year.to = 2100, statusMin = 3,
  df.1 = df.1)
}
\arguments{
\item{year.from}{Starting year}

\item{year.to}{Ending year}

\item{statusMin}{Minimum status of seizure record to include in summary (usually 3)}

\item{df.1}{Dataframe with list of seizure records to be considered.
Created using df_quantities_RIE_separate.fn}
}
\value{
A list with four items.
Each item is a dataframe with information about the seizures
for which there are multiple countries of origin. The dataframes are:
\tabular{ll}{
disc \tab Country of discovery \cr
 \tab One row for each seizure \cr
transit \tab Trade route countries \cr
 \tab Multiple rows for each seizure to give all combinations
         of export and transit countries (with seizure opportunity) \cr
orig_raw \tab Countries of origin for raw ivory \cr
\tab One row for each country of origin for each seizure.
         Includes the proportion of shipment allocated to each country \cr
orig_wkd \tab Countries of origin for worked ivory\cr
 \tab One row for each country of origin for each seizure.
         Includes the proportion of shipment allocated to each country \cr
}
}
\description{
The weight of a seizure with multiple countries of origin
is allocated according to the proportion allocated to each country.
It also identifies the countries listed at each stage in the trade chain.
This is required for trade flows and le ratio accounting.
This function interacts with the ETIS database and the seizures
stored in df.1 created by df_quantities_RIE_separate.fn
to identify all seizures with multiple countries of origin.
For each seizure with multiple countries of origin it extracts
countries of origin, discovery and transit (including export).
The main use of this function is to contribute to the calculation
of two new database tables for law enforcement ratio and trade flow accounting.
}
\examples{
year.from <- 1900
year.to <- 2100
statusMin <- 3

df.1 <- df_quantities_RIE_separate(year.from = year.from, year.to = year.to,
statusMin = statusMin, reg.model = 'wt est.Rdata')

mult.dat <- mult_ctries(year.from = 1900, year.to = 2100, statusMin = 3, df.1 = df.1)
}
\seealso{
\code{\link{df_quantities_RIE_separate}},
\code{\link{inout_tables_LE_TF_prep}},
\code{\link{inout_tables_LE_TF_final}}
}
